refactor(codegen): extrai passes/object_methods.rs de func.rs
Move desugar_object_methods (235 linhas) para
crates/rts-codegen/src/codegen/lower/passes/object_methods.rs.
Reescreve `{ foo() {...} }` em `{ foo: function() {...} }` antes
do hoist_fn_expressions.

func.rs: 7413 -> 7178 linhas. Suite verde.
