+++ crates/rts-codegen/src/codegen/lower/func.rs
@@ -29,6 +29,7 @@ use super::analysis::types::sanitize_symbol;
 use super::passes::args::default_args::expand_default_args;
 use super::passes::args::rest_args::expand_rest_args;
 use super::passes::args::spread_args::expand_spread_args;
+use super::passes::object_methods::desugar_object_methods;
 use super::ctx::{ClassMeta, FnCtx, GlobalVar, UserFnAbi, ValTy};
 use super::statements::lower_stmt;
 
@@ -2655,244 +2656,6 @@ fn expand_async_functions(program: &mut Program) {
 /// (try_lower_fn_expr_decl); este pass cobre o que sobrou. Nao tenta
 /// resolver capturas de escopo — se o body referenciar variaveis
 /// fora dele, falha em codegen como ident undefined.
-/// (#261) Desugara \`{ method() { body } }\` em \`{ method: function() { body } }\`
-/// antes do hoist_fn_expressions. Permite que methods em object literal sejam
-/// hoisted como user fns top-level e despachados como handles.
-/// \`this\` no body funciona via thread-local slot (PR #451) — o call site
-/// \`obj.method()\` empilha obj antes do invoke.
-fn desugar_object_methods(program: &mut Program) {
-    use swc_ecma_ast::{Prop, PropOrSpread};
-
-    fn visit_expr(expr: &mut Expr) {
-        match expr {
-            Expr::Object(obj) => {
-                for prop in obj.props.iter_mut() {
-                    if let PropOrSpread::Prop(p) = prop {
-                        // Visita sub-exprs em KeyValue antes de tentar desugar.
-                        if let Prop::KeyValue(kv) = p.as_mut() {
-                            visit_expr(kv.value.as_mut());
-                            continue;
-                        }
-                        // Desugar Method para KeyValue + Fn.
-                        if matches!(p.as_ref(), Prop::Method(_)) {
-                            let owned = std::mem::replace(
-                                p.as_mut(),
-                                Prop::KeyValue(swc_ecma_ast::KeyValueProp {
-                                    key: swc_ecma_ast::PropName::Ident(
-                                        swc_ecma_ast::IdentName::new("__placeholder".into(), Default::default()),
-                                    ),
-                                    value: Box::new(Expr::Invalid(swc_ecma_ast::Invalid {
-                                        span: Default::default(),
-                                    })),
-                                }),
-                            );
-                            if let Prop::Method(method) = owned {
-                                let mut function = method.function;
-                                if let Some(body) = function.body.as_mut() {
-                                    for stmt in body.stmts.iter_mut() {
-                                        visit_stmt(stmt);
-                                    }
-                                }
-                                let fn_expr = swc_ecma_ast::FnExpr {
-                                    ident: None,
-                                    function,
-                                };
-                                **p = Prop::KeyValue(swc_ecma_ast::KeyValueProp {
-                                    key: method.key,
-                                    value: Box::new(Expr::Fn(fn_expr)),
-                                });
-                            }
-                        }
-                    }
-                }
-            }
-            Expr::Array(arr) => {
-                for elem in arr.elems.iter_mut().flatten() {
-                    visit_expr(elem.expr.as_mut());
-                }
-            }
-            Expr::Call(c) => {
-                if let swc_ecma_ast::Callee::Expr(callee) = &mut c.callee {
-                    visit_expr(callee.as_mut());
-                }
-                for arg in c.args.iter_mut() {
-                    visit_expr(arg.expr.as_mut());
-                }
-            }
-            Expr::New(n) => {
-                visit_expr(n.callee.as_mut());
-                if let Some(args) = n.args.as_mut() {
-                    for arg in args.iter_mut() {
-                        visit_expr(arg.expr.as_mut());
-                    }
-                }
-            }
-            Expr::Bin(b) => {
-                visit_expr(b.left.as_mut());
-                visit_expr(b.right.as_mut());
-            }
-            Expr::Assign(a) => visit_expr(a.right.as_mut()),
-            Expr::Cond(c) => {
-                visit_expr(c.test.as_mut());
-                visit_expr(c.cons.as_mut());
-                visit_expr(c.alt.as_mut());
-            }
-            Expr::Member(m) => visit_expr(m.obj.as_mut()),
-            Expr::Paren(p) => visit_expr(p.expr.as_mut()),
-            Expr::TsAs(a) => visit_expr(a.expr.as_mut()),
-            Expr::TsTypeAssertion(a) => visit_expr(a.expr.as_mut()),
-            Expr::TsConstAssertion(a) => visit_expr(a.expr.as_mut()),
-            Expr::TsSatisfies(a) => visit_expr(a.expr.as_mut()),
-            Expr::TsNonNull(n) => visit_expr(n.expr.as_mut()),
-            Expr::Await(a) => visit_expr(a.arg.as_mut()),
-            Expr::Tpl(t) => {
-                for e in t.exprs.iter_mut() {
-                    visit_expr(e.as_mut());
-                }
-            }
-            Expr::Unary(u) => visit_expr(u.arg.as_mut()),
-            Expr::Update(u) => visit_expr(u.arg.as_mut()),
-            Expr::Seq(s) => {
-                for e in s.exprs.iter_mut() {
-                    visit_expr(e.as_mut());
-                }
-            }
-            Expr::Fn(f) => {
-                if let Some(body) = f.function.body.as_mut() {
-                    for stmt in body.stmts.iter_mut() {
-                        visit_stmt(stmt);
-                    }
-                }
-            }
-            Expr::Arrow(arrow) => match arrow.body.as_mut() {
-                swc_ecma_ast::BlockStmtOrExpr::BlockStmt(b) => {
-                    for stmt in b.stmts.iter_mut() {
-                        visit_stmt(stmt);
-                    }
-                }
-                swc_ecma_ast::BlockStmtOrExpr::Expr(e) => visit_expr(e.as_mut()),
-            },
-            _ => {}
-        }
-    }
-
-    fn visit_stmt(stmt: &mut Stmt) {
-        match stmt {
-            Stmt::Expr(e) => visit_expr(e.expr.as_mut()),
-            Stmt::Decl(swc_ecma_ast::Decl::Var(v)) => {
-                for d in v.decls.iter_mut() {
-                    if let Some(init) = &mut d.init {
-                        visit_expr(init.as_mut());
-                    }
-                }
-            }
-            Stmt::Return(r) => {
-                if let Some(arg) = &mut r.arg {
-                    visit_expr(arg.as_mut());
-                }
-            }
-            Stmt::If(i) => {
-                visit_expr(i.test.as_mut());
-                visit_stmt(i.cons.as_mut());
-                if let Some(alt) = &mut i.alt {
-                    visit_stmt(alt.as_mut());
-                }
-            }
-            Stmt::Block(b) => {
-                for s in b.stmts.iter_mut() {
-                    visit_stmt(s);
-                }
-            }
-            Stmt::While(w) => {
-                visit_expr(w.test.as_mut());
-                visit_stmt(w.body.as_mut());
-            }
-            Stmt::DoWhile(w) => {
-                visit_expr(w.test.as_mut());
-                visit_stmt(w.body.as_mut());
-            }
-            Stmt::For(f) => {
-                if let Some(init) = &mut f.init {
-                    match init {
-                        swc_ecma_ast::VarDeclOrExpr::VarDecl(v) => {
-                            for d in v.decls.iter_mut() {
-                                if let Some(i) = &mut d.init {
-                                    visit_expr(i.as_mut());
-                                }
-                            }
-                        }
-                        swc_ecma_ast::VarDeclOrExpr::Expr(e) => visit_expr(e.as_mut()),
-                    }
-                }
-                if let Some(t) = &mut f.test {
-                    visit_expr(t.as_mut());
-                }
-                if let Some(u) = &mut f.update {
-                    visit_expr(u.as_mut());
-                }
-                visit_stmt(f.body.as_mut());
-            }
-            Stmt::ForOf(f) => {
-                visit_expr(f.right.as_mut());
-                visit_stmt(f.body.as_mut());
-            }
-            Stmt::ForIn(f) => {
-                visit_expr(f.right.as_mut());
-                visit_stmt(f.body.as_mut());
-            }
-            Stmt::Throw(t) => visit_expr(t.arg.as_mut()),
-            Stmt::Try(t) => {
-                for s in t.block.stmts.iter_mut() {
-                    visit_stmt(s);
-                }
-                if let Some(handler) = &mut t.handler {
-                    for s in handler.body.stmts.iter_mut() {
-                        visit_stmt(s);
-                    }
-                }
-                if let Some(finalizer) = &mut t.finalizer {
-                    for s in finalizer.stmts.iter_mut() {
-                        visit_stmt(s);
-                    }
-                }
-            }
-            Stmt::Switch(s) => {
-                visit_expr(s.discriminant.as_mut());
-                for case in s.cases.iter_mut() {
-                    if let Some(test) = &mut case.test {
-                        visit_expr(test.as_mut());
-                    }
-                    for stmt in case.cons.iter_mut() {
-                        visit_stmt(stmt);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-
-    for item in program.items.iter_mut() {
-        match item {
-            crate::parser::ast::Item::Function(fdecl) => {
-                for stmt in fdecl.body.iter_mut() {
-                    let crate::parser::ast::Statement::Raw(raw) = stmt;
-                    if let Some(s) = raw.stmt.as_mut() {
-                        visit_stmt(s);
-                    }
-                }
-            }
-            crate::parser::ast::Item::Class(_) => {}
-            crate::parser::ast::Item::Statement(stmt) => {
-                let crate::parser::ast::Statement::Raw(raw) = stmt;
-                if let Some(s) = raw.stmt.as_mut() {
-                    visit_stmt(s);
-                }
-            }
-            _ => {}
-        }
-    }
-}
-
 fn hoist_fn_expressions(program: &mut Program) {
     use swc_ecma_ast::{ArrowExpr, BlockStmtOrExpr};
